------------------------ Fast BER coder prototype ------------------------
These additions and changes constitute a prototype for fast BER
encoding and decoding.  The conventional coders rely on Java
reflection heavily, which severely limits the speed of their
execution.  To eliminate as much reflection as possible, I've:

  - Created a set of ordinary Java classes, org.bn.metadata.*, that
    encapsulate the information currently represented with
    annotations.

  - Changed the BNCompiler code generator to produce static members
    containing this metadata along with statically initialized
    references to Fields that were previously fetched via Java
    reflection.

  - Written FastDecoder and FastBERDecoder classes that exploit the
    aforementioned static metadata.

In my test, I've seen decode times improve from around 200/sec to over
17K/sec.

Note that to a large extent, this prototype is tailored to my test
case, and thus doesn't handle all of the ASN.1 constructs that the
existing coders do.

To do:
------

  - Complete the FastEncoder and FastBEREncoder implementations.
  - Add support for missing ASN.1 constructs.